ingestion_api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
pydantic
python-dotenv
google-cloud-secret-manager
requests
